+++ src/bindings.c
@@ -91,7 +91,7 @@ int vq_ext_devinfo(int handle, int device, int *dev_exists, char *file_path, cha
 	Erweiterter v_opnwk, mit dem auch die erweiterten Parameter
 	von NVDI unterstÅtzt werden.
 	-----------------------------------------------------------*/
-void v_ext_opnwk(int *work_in, int *handle, int *work_out)
+void v_ext_opnwk(WORD *work_in, WORD *handle, WORD *work_out)
 {
 	VDIPB vdipb;
 
@@ -275,7 +275,7 @@ int v_delete_ctab(int handle, COLOR_TAB *ctab)
 }
 
 
-int	objc_sysvar(int ob_smode, int ob_swhich, int ob_sival1, int ob_sival2, int *ob_soval1, int *ob_soval2)
+WORD	objc_sysvar(WORD ob_smode, WORD ob_swhich, WORD ob_sival1, WORD ob_sival2, WORD *ob_soval1, WORD *ob_soval2)
 {
 	AESPB aespb;
